+++ climate-data-ca/resources/js/analyze-functions.js
@@ -927,6 +927,7 @@
 
             var submit_data = {
                 'analyze-captcha_code': $('#analyze-captcha_code').val(),
+                "signup":$('#signup').is(":checked"),
                 'request_data': form_obj,
                 'submit_url': submit_url_var + submit_url_post // ex: wetdays/jobs
             }