+++ climate-data-ca/template/download/variable.php
@@ -213,7 +213,10 @@
                             <a class="btn btn-secondary text-white all-caps download-process-btn disabled" id="daily-process" target="_blank"><?php _e('Send Request', 'cdc'); ?> <i class="far fa-arrow-alt-circle-down"></i></a>
                         </div>
                     </div>
-
+                    <div class="form-check form-check-inline mt-4">
+                        <input class="form-check-input" type="checkbox" value="" id="signup">
+                        <label class="form-check-label" for="signup"><?php _e ( 'Subscribe to the ClimateData.ca newsletter', 'cdc' ); ?></label>
+                    </div>
                 </div>
 
                 <div class="col-8-of-8 p-4">