+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+2022-03-10.01
+  * Added mailchimp subscribe integration to analyze, download and feedback form
+
 2022-01-20.01
   * Added netcdf format for variable download